+++ proyecto sistema operativos 2/src/main/java/com/umg/backend/repository/BitacoraRepository.java
@@ -4,7 +4,7 @@ package com.umg.backend.repository;
 import com.umg.backend.entity.Bitacora;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BitacoraRepository extends JpaRepository<Bitacora, Long> {
+public interface BitacoraRepository extends JpaRepository<Bitacora, Integer> {
 
 
 }
