+++ proyecto sistema operativos 2/src/main/java/com/umg/backend/repository/EmpleadoRepository.java
@@ -3,6 +3,8 @@ package com.umg.backend.repository;
 import com.umg.backend.entity.Empleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmpleadoRepository extends JpaRepository <Empleado, Long> {
+
+public interface EmpleadoRepository extends JpaRepository <Empleado, Integer> {
+
 
 }
