+++ proyecto sistema operativos 2/src/main/java/com/umg/backend/repository/PersonaRepository.java
@@ -3,5 +3,5 @@ package com.umg.backend.repository;
 import com.umg.backend.entity.Persona;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonaRepository extends JpaRepository<Persona, Long> {
+public interface PersonaRepository extends JpaRepository<Persona, Integer> {
 }
