+++ proyecto sistema operativos 2/src/main/java/com/umg/backend/repository/PuestoEmpleadoRepository.java
@@ -5,5 +5,5 @@ package com.umg.backend.repository;
 import com.umg.backend.entity.PuestoEmpleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PuestoEmpleadoRepository extends JpaRepository<PuestoEmpleado, Long> {
+public interface PuestoEmpleadoRepository extends JpaRepository<PuestoEmpleado, Integer> {
 }
