+++ proyecto sistema operativos 2/src/main/java/com/umg/backend/repository/PuestoRepository.java
@@ -4,5 +4,5 @@ package com.umg.backend.repository;
 import com.umg.backend.entity.Puesto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PuestoRepository extends JpaRepository<Puesto, Long> {
+public interface PuestoRepository extends JpaRepository<Puesto, Integer> {
 }
